+++ warp/tests/test_quat.py
@@ -1886,7 +1886,7 @@ def test_quat_identity(test, device, dtype, register_kernels=False):
 ############################################################
 
 
-def test_quat_assign(test, device, dtype, register_kernels=False):
+def test_quat_assign_inplace(test, device, dtype, register_kernels=False):
     np_type = np.dtype(dtype)
     wp_type = wp.types.np_dtype_to_warp_type[np_type]
 
@@ -1907,16 +1907,6 @@ def test_quat_assign(test, device, dtype, register_kernels=False):
         g = q[0] + wp_type(2.0) * q[1] + wp_type(3.0) * q[2] + wp_type(4.0) * q[3]
         x[tid] = g
 
-    def quattest_in_register_overwrite(x: wp.array(dtype=quat), a: wp.array(dtype=quat)):
-        tid = wp.tid()
-
-        f = quat()
-        a_quat = a[tid]
-        f = a_quat
-        f[1] = wp_type(3.0)
-
-        x[tid] = f
-
     def quattest_component(x: wp.array(dtype=quat), y: wp.array(dtype=wp_type)):
         i = wp.tid()
 
@@ -1929,7 +1919,6 @@ def test_quat_assign(test, device, dtype, register_kernels=False):
 
     kernel_read_write_store = getkernel(quattest_read_write_store, suffix=dtype.__name__)
     kernel_in_register = getkernel(quattest_in_register, suffix=dtype.__name__)
-    kernel_in_register_overwrite = getkernel(quattest_in_register_overwrite, suffix=dtype.__name__)
     kernel_component = getkernel(quattest_component, suffix=dtype.__name__)
 
     if register_kernels:
@@ -1965,7 +1954,6 @@ def test_quat_assign(test, device, dtype, register_kernels=False):
     x = wp.zeros(1, dtype=quat, requires_grad=True)
     y = wp.ones(1, dtype=wp_type, requires_grad=True)
 
-    tape = wp.Tape()
     with tape:
         wp.launch(kernel_component, dim=1, inputs=[x, y])
 
@@ -1973,18 +1961,6 @@ def test_quat_assign(test, device, dtype, register_kernels=False):
 
     assert_np_equal(x.numpy(), np.array([[1.0, 2.0, 3.0, 4.0]], dtype=np_type))
     assert_np_equal(y.grad.numpy(), np.array([10.0], dtype=np_type))
-
-    x = wp.zeros(1, dtype=quat, device=device, requires_grad=True)
-    a = wp.ones(1, dtype=quat, device=device, requires_grad=True)
-
-    tape = wp.Tape()
-    with tape:
-        wp.launch(kernel_in_register_overwrite, dim=1, inputs=[x, a], device=device)
-
-    tape.backward(grads={x: wp.ones_like(x, requires_grad=False)})
-
-    assert_np_equal(x.numpy(), np.array([[1.0, 3.0, 1.0, 1.0]], dtype=np_type))
-    assert_np_equal(a.grad.numpy(), np.array([[1.0, 0.0, 1.0, 1.0]], dtype=np_type))
 
 
 ############################################################
@@ -2120,7 +2096,7 @@ def test_quat_len(test, device):
 
 
 @wp.kernel
-def vector_augassign_kernel(
+def quat_augassign_kernel(
     a: wp.array(dtype=wp.quat), b: wp.array(dtype=wp.quat), c: wp.array(dtype=wp.quat), d: wp.array(dtype=wp.quat)
 ):
     i = wp.tid()
@@ -2138,26 +2114,26 @@ def vector_augassign_kernel(
     q3 = wp.quat()
     q4 = d[i]
 
-    q3[0] += q4[0]
-    q3[1] += q4[1]
-    q3[2] += q4[2]
-    q3[3] += q4[3]
+    q3[0] -= q4[0]
+    q3[1] -= q4[1]
+    q3[2] -= q4[2]
+    q3[3] -= q4[3]
 
-    c[i] = q1
+    c[i] = q3
 
 
-def test_vector_augassign(test, device):
+def test_quat_augassign(test, device):
     N = 3
 
-    a = wp.zeros(N, dtype=wp.quat, requires_grad=True)
-    b = wp.ones(N, dtype=wp.quat, requires_grad=True)
+    a = wp.zeros(N, dtype=wp.quat, requires_grad=True, device=device)
+    b = wp.ones(N, dtype=wp.quat, requires_grad=True, device=device)
 
-    c = wp.zeros(N, dtype=wp.quat, requires_grad=True)
-    d = wp.ones(N, dtype=wp.quat, requires_grad=True)
+    c = wp.zeros(N, dtype=wp.quat, requires_grad=True, device=device)
+    d = wp.ones(N, dtype=wp.quat, requires_grad=True, device=device)
 
     tape = wp.Tape()
     with tape:
-        wp.launch(vector_augassign_kernel, N, inputs=[a, b, c, d])
+        wp.launch(quat_augassign_kernel, N, inputs=[a, b, c, d], device=device)
 
     tape.backward(grads={a: wp.ones_like(a), c: wp.ones_like(c)})
 
@@ -2168,6 +2144,38 @@ def test_vector_augassign(test, device):
     assert_np_equal(c.numpy(), -wp.ones_like(c).numpy())
     assert_np_equal(c.grad.numpy(), wp.ones_like(c).numpy())
     assert_np_equal(d.grad.numpy(), -wp.ones_like(d).numpy())
+
+
+def test_quat_assign_copy(test, device):
+    saved_enable_vector_component_overwrites_setting = wp.config.enable_vector_component_overwrites
+    try:
+        wp.config.enable_vector_component_overwrites = True
+
+        @wp.kernel
+        def quat_in_register_overwrite(x: wp.array(dtype=wp.quat), a: wp.array(dtype=wp.quat)):
+            tid = wp.tid()
+
+            f = wp.quat()
+            a_quat = a[tid]
+            f = a_quat
+            f[1] = 3.0
+
+            x[tid] = f
+
+        x = wp.zeros(1, dtype=wp.quat, device=device, requires_grad=True)
+        a = wp.ones(1, dtype=wp.quat, device=device, requires_grad=True)
+
+        tape = wp.Tape()
+        with tape:
+            wp.launch(quat_in_register_overwrite, dim=1, inputs=[x, a], device=device)
+
+        tape.backward(grads={x: wp.ones_like(x, requires_grad=False)})
+
+        assert_np_equal(x.numpy(), np.array([[1.0, 3.0, 1.0, 1.0]], dtype=float))
+        assert_np_equal(a.grad.numpy(), np.array([[1.0, 0.0, 1.0, 1.0]], dtype=float))
+
+    finally:
+        wp.config.enable_vector_component_overwrites = saved_enable_vector_component_overwrites_setting
 
 
 devices = get_test_devices()
@@ -2269,8 +2277,8 @@ for dtype in np_float_types:
     )
     add_function_test_register_kernel(
         TestQuat,
-        f"test_quat_assign_{dtype.__name__}",
-        test_quat_assign,
+        f"test_quat_assign_inplace_{dtype.__name__}",
+        test_quat_assign_inplace,
         devices=devices,
         dtype=dtype,
     )
@@ -2279,7 +2287,8 @@ for dtype in np_float_types:
     )
 
 add_function_test(TestQuat, "test_quat_len", test_quat_len, devices=devices)
-
+add_function_test(TestQuat, "test_quat_augassign", test_quat_augassign, devices=devices)
+add_function_test(TestQuat, "test_quat_assign_copy", test_quat_assign_copy, devices=devices)
 
 if __name__ == "__main__":
     wp.clear_kernel_cache()