+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - `warp.fem.integrate()` and `warp.fem.integrate()` may now perform parallel evaluation of quadrature points within elements
 - `warp.fem.interpolate()` can now build Jacobian sparse matrices of interpolated functions with respect to a trial field
+- Vector/matrix/quaternion component assignment operations compile and run faster in the backward pass. Assignment should only happen once per component.
 
 ### Fixed